move lib types to new src/lib/types.ts
