+++ src/lib/index.ts
@@ -2,49 +2,7 @@ export { default as CircleSpinner } from './CircleSpinner.svelte'
 export { default as CmdPalette } from './CmdPalette.svelte'
 export { default as MultiSelect, default } from './MultiSelect.svelte'
 export { default as Wiggle } from './Wiggle.svelte'
-
-export type Option = string | number | ObjectOption
-
-export type ObjectOption = {
-  label: string | number // user-displayed text
-  value?: unknown // associated value, can be anything incl. objects (defaults to label if undefined)
-  title?: string // on-hover tooltip
-  disabled?: boolean // make this option unselectable
-  preselected?: boolean // make this option selected on page load (before any user interaction)
-  disabledTitle?: string // override the default disabledTitle = 'This option is disabled'
-  selectedTitle?: string // tooltip to display when this option is selected and hovered
-  [key: string]: unknown // allow any other keys users might want
-}
-
-export type DispatchEvents<T = Option> = {
-  add: { option: T }
-  create: { option: T }
-  remove: { option: T }
-  removeAll: { options: T[] }
-  change: {
-    option?: T
-    options?: T[]
-    type: 'add' | 'remove' | 'removeAll'
-  }
-  open: { event: Event }
-  close: { event: Event }
-}
-
-export type MultiSelectEvents = {
-  [key in keyof DispatchEvents]: CustomEvent<DispatchEvents[key]>
-} & {
-  blur: FocusEvent
-  click: MouseEvent
-  focus: FocusEvent
-  keydown: KeyboardEvent
-  keyup: KeyboardEvent
-  mouseenter: MouseEvent
-  mouseleave: MouseEvent
-  touchcancel: TouchEvent
-  touchend: TouchEvent
-  touchmove: TouchEvent
-  touchstart: TouchEvent
-}
+export * from './types'
 
 // Firefox lacks support for scrollIntoViewIfNeeded (https://caniuse.com/scrollintoviewifneeded).
 // See https://github.com/janosh/svelte-multiselect/issues/87
